requirements.txt
tensorflow_gpu==1.12.0
numpy==1.16.4
opencv_python==4.1.1.26